IHC/ScrumBot/requirements.txt
gradio
requests
fpdf
PyMuPDF
duckduckgo-search
python-dotenv